Move initializer in each page to index.js
